# swagger/doc/cms.json
{
	"swagger": "2.0",
	"info": {
		"title": "Cozrum",
		"version": "1",
		"description": "Response format: #definitions/Response"
	},
	"basePath": "/api/v1",
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"security": [
		{
			"JWT": []
		}
	],
	"paths": {
...
		"/user/changePassword": {
			"post": {
				"tags": [
					"user"
				],
				"operationId": "changePassword",
				"summary": "Đổi mật khẩu",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "Account",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"oldPassword": {
									"type": "string",
									"default": "admin"
								},
								"newPassword": {
									"type": "string",
									"default": "admin"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
		"/user/resetPassword/{userId}": {
			"post": {
				"tags": [
					"user"
				],
				"operationId": "resetPassword",
				"summary": "Khôi phục mật khẩu",
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"description": "User ID",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"newPassword": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
...
		"/account": {
			"get": {
				"tags": [
					"account"
				],
				"operationId": "getAccounts",
				"summary": "Danh sách accounts",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "start",
						"description": "Start",
						"type": "number"
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Limit",
						"type": "number"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"put": {
				"tags": [
					"account"
				],
				"operationId": "addAccount",
				"summary": "Thêm account",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "RoleGroup",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
		"/account/changRole/{userId}": {
			"put": {
				"tags": [
					"account"
				],
				"operationId": "changeRole",
				"summary": "Thay đổi quyền user",
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"description": "UserId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "RoleGroup",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"role": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
		"/account/lock/{userId}": {
			"post": {
				"tags": [
					"account"
				],
				"operationId": "changeRole",
				"summary": "Lock / unlock account",
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"description": "UserId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"enable": {
									"type": "boolean"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
		"/role/": {
			"get": {
				"tags": [
					"role"
				],
				"operationId": "getRoleGroup",
				"summary": "Lấy thông tin nhóm quyền",
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "object",
							"properties": {
								"roles": {
									"type": "array",
									"items": {
										"$ref": "#/definitions/RoleGroup"
									}
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"role"
				],
				"operationId": "addRoleGroup",
				"summary": "Thêm nhóm quyền",
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "RoleGroup",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RoleGroup"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/RoleGroup"
						}
					}
				}
			}
		},
		"/role/{roleName}": {
			"put": {
				"tags": [
					"role"
				],
				"operationId": "updateRoleGroup",
				"summary": "Cập nhật nhóm quyền",
				"parameters": [
					{
						"in": "path",
						"name": "roleName",
						"description": "Role Name",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "RoleGroup",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RoleGroup"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/RoleGroup"
						}
					}
				}
			}
		},
...
		"/ota/account": {
			"get": {
				"tags": [
					"ota"
				],
				"operationId": "otaAccount",
				"summary": "Lấy tên các tài khoản OTA",
				"produces": [
					"application/json"
				],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"schema": {
							"type": "object",
							"properties": {
								"accounts": {
									"type": "array",
									"items": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
		"/setting/crawler": {
			"post": {
				"tags": [
					"setting"
				],
				"operationId": "crawler",
				"summary": "crawler OTA",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "Auto data",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"otas": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "settings:"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		},
...
	},
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"in": "header",
			"name": "x-access-token"
		}
	},
	"definitions": {
		"User": {
			"type": "object",
			"properties": {
				"username": {
					"type": "string"
				},
				"passwords": {
					"type": "string"
				},
				"group": {
					"type": "string"
				},
				"enable": {
					"type": "boolean"
				},
				"name": {
					"type": "string"
				},
				"city": {
					"type": "string"
				},
				"address": {
					"type": "string"
				},
				"phone": {
					"type": "string"
				}
			}
		},
...
		"RoleGroup": {
			"type": "object",
			"properties": {
				"role": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"groups": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"name": {
								"type": "string"
							},
							"methods": {
								"type": "array",
								"items": {
									"type": "string",
									"enum": [
										"GET",
										"POST",
										"PUT",
										"DELETE",
										"*"
									]
								}
							}
						}
					}
				},
				"tabs": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
...
	}
}